Move types their logical modules
