+++ src/lib.rs
@@ -3,4 +3,3 @@ pub mod decoder;
 pub mod droplet;
 pub mod encoder;
 pub mod soliton;
-pub mod types;
