+++ tests/ideal_enc_dec_prop_test.rs
@@ -1,7 +1,6 @@
 use fountaincode::{
-    decoder::Decoder,
-    encoder::Encoder,
-    types::{CatchResult, EncoderType},
+    decoder::{CatchResult, Decoder},
+    encoder::{Encoder, EncoderType},
 };
 use proptest::prelude::*;
 use rand::{distributions::Alphanumeric, thread_rng, Rng};
